// apps/web/src/types/components/graphql.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Date: { input: any; output: any; }
  ObjectId: { input: any; output: any; }
};

export type Cpu = {
  __typename?: 'CPU';
  cooler: Scalars['Boolean']['output'];
  coreClock: Scalars['String']['output'];
  coreCount: Scalars['String']['output'];
  coreFamily: Scalars['String']['output'];
  id: Scalars['ObjectId']['output'];
  integratedGraphics: Scalars['String']['output'];
  manufacturer: Scalars['String']['output'];
  microarchitecture: Scalars['String']['output'];
  name: Scalars['String']['output'];
  packaging: PackagingType;
  partNumber: Scalars['String']['output'];
  price: Scalars['String']['output'];
  series: Scalars['String']['output'];
  socket: Scalars['String']['output'];
  tdp: Scalars['String']['output'];
};

export type Case = {
  __typename?: 'Case';
  color: Scalars['String']['output'];
  componentType: ComponentType;
  formFactor: Scalars['String']['output'];
  id: Scalars['String']['output'];
  interface: Scalars['String']['output'];
  manufacturer: Scalars['String']['output'];
  name: Scalars['String']['output'];
  partNumber: Scalars['String']['output'];
  powerSupply: Scalars['Boolean']['output'];
  price: Scalars['String']['output'];
  sidePanel: SidePanelType;
  type: CaseType;
};

export enum CaseType {
  Atx = 'ATX',
  AtxMidTower = 'ATX_MID_TOWER',
  Eatx = 'EATX',
  MicroAtx = 'MICRO_ATX',
  MiniItx = 'MINI_ITX'
}

export enum ComponentType {
  Case = 'CASE',
  Cpu = 'CPU',
  Memory = 'MEMORY',
  Motherboard = 'MOTHERBOARD',
  PowerSupply = 'POWER_SUPPLY',
  Storage = 'STORAGE',
  VideoCard = 'VIDEO_CARD'
}
// ...
export type Memory = {
  __typename?: 'Memory';
  color: Scalars['String']['output'];
  formFactor: Scalars['String']['output'];
  heatSpreader: Scalars['Boolean']['output'];
  id: Scalars['ObjectId']['output'];
  manufacturer: Scalars['String']['output'];
  modules: Scalars['String']['output'];
  name: Scalars['String']['output'];
  partNumber: Scalars['String']['output'];
  price: Scalars['String']['output'];
  speed: Scalars['String']['output'];
  voltage: Scalars['String']['output'];
};

export enum MemoryType {
  Ddr4 = 'DDR4',
  Ddr5 = 'DDR5'
}

export type Motherboard = {
  __typename?: 'Motherboard';
  chipset: Scalars['String']['output'];
  color: Scalars['String']['output'];
  formFactor: Scalars['String']['output'];
  id: Scalars['ObjectId']['output'];
  manufacturer: Scalars['String']['output'];
  memoryMax: Scalars['String']['output'];
  memorySlots: Scalars['String']['output'];
  memoryType: MemoryType;
  name: Scalars['String']['output'];
  partNumber: Scalars['String']['output'];
  price: Scalars['String']['output'];
  socket: Scalars['String']['output'];
};
// ...
export type PowerSupply = {
  __typename?: 'PowerSupply';
  SATAConnectors: Scalars['String']['output'];
  color: Scalars['String']['output'];
  fanless: Scalars['Boolean']['output'];
  id: Scalars['ObjectId']['output'];
  length: Scalars['String']['output'];
  manufacturer: Scalars['String']['output'];
  model: Scalars['String']['output'];
  name: Scalars['String']['output'];
  partNumber: Scalars['String']['output'];
  price: Scalars['String']['output'];
  type: Scalars['String']['output'];
  wattage: Scalars['String']['output'];
};
// ...
export enum SidePanelType {
  TemperedGlass = 'TEMPERED_GLASS',
  TintedTemperedGlass = 'TINTED_TEMPERED_GLASS'
}

export type Storage = {
  __typename?: 'Storage';
  NVME: Scalars['Boolean']['output'];
  capacity: Scalars['String']['output'];
  formFactor: Scalars['String']['output'];
  id: Scalars['ObjectId']['output'];
  interface: Scalars['String']['output'];
  manufacturer: Scalars['String']['output'];
  name: Scalars['String']['output'];
  partNumber: Scalars['String']['output'];
  price: Scalars['String']['output'];
  type: StorageType;
};
// ...
export type VideoCard = {
  __typename?: 'VideoCard';
  HDMIOutputs: Scalars['String']['output'];
  TDP: Scalars['String']['output'];
  chipset: Scalars['String']['output'];
  color: Scalars['String']['output'];
  coolingFans: Scalars['String']['output'];
  coreClock: Scalars['String']['output'];
  displayPortOutputs: Scalars['String']['output'];
  id: Scalars['ObjectId']['output'];
  interface: Scalars['String']['output'];
  manufacturer: Scalars['String']['output'];
  memory: Scalars['String']['output'];
  memoryType: Scalars['String']['output'];
  model: Scalars['String']['output'];
  name: Scalars['String']['output'];
  partNumber: Scalars['String']['output'];
  price: Scalars['String']['output'];
};
// ...
    interface
    powerSupply
    sidePanel